Revise workaround for dealing with wpautop shortcode mangling
* Parse post content for shortcodes at filter priority 9 (before wpautop runs), replace all shortcode content with a random string, then drop shortcode content back into content at filter priority 10 (after wpautop, before do_shortcode). The previous workaround, involving changing the priority of wpautop, had the potential to break other plugins implementing their own workaround.
* Refactor functions.php to isolate the (now much larger) shortcodes functions into their own file.
