+++ functions.php
@@ -5,33 +5,6 @@ add_theme_support( 'title-tag' );
 add_theme_support( 'automatic-feed-links' );
 add_theme_support( 'post-thumbnails' );
 
-// This is bad, but necessary. wpautop(), apart from being a poor idea in the first place (IMO),
-// doesn't know anything about shortcodes and consistently breaks shortcodes with content. The
-// shortcode_unautop() function is supposed to undo the damage, but it uses regular expressions
-// like a hammer to solve a problem which in reality needs a token parser (nested tags) and
-// therefore does not work.
-//
-// That leaves the hapless theme/plugin developer three options:
-//
-// 1. Rejigger the shortcode filter priority to bump it ahead of wpautop(). This could break
-//    other plugins relying on priority.
-// 2. Disable wpautop() entirely. (I would rejoice in principle, but most other people like wpautop.)
-// 3. Leave shortcodes broken and generating invalid output, and just recommend to users that they
-//    download a plugin to disable wpautop on a per-post basis.
-//
-// I'm taking option #1 here as the lesser evil. wpautop() still puts extraneous <p> elements all
-// over the dang place, but at least it no longer walks all over the shortcode-generated markup.
-//
-// For more information about this problem, see also:
-//
-// - https://core.trac.wordpress.org/ticket/6984
-// - https://core.trac.wordpress.org/ticket/14050
-// - http://customcreative.co.uk/resolving-wpautop-and-shortcodes/
-// - http://betterwp.net/protect-shortcodes-from-wpautop-and-the-likes/
-
-remove_filter( 'the_content', 'do_shortcode' );
-add_filter( 'the_content', 'do_shortcode', 9 );
-
 // Enqueue main styles
 add_action( 'wp_enqueue_scripts', function() {
 	global $wp_styles;
@@ -115,122 +88,7 @@ function woocommerce_support() {
 // Shortcodes
 // ==========================================================
 
-add_shortcode( 'columns', 'simplecommerce_shortcode_columns' );
-add_shortcode( 'column', 'simplecommerce_shortcode_column' );
-add_shortcode( 'testimonial', 'simplecommerce_shortcode_testimonial' );
-add_shortcode( 'toggle', 'simplecommerce_shortcode_toggle' );
-add_shortcode( 'contentbox', 'simplecommerce_shortcode_contentbox' );
-add_filter( 'no_texturize_shortcodes', function( $non_texturized_shortcodes ) {
-	$non_texturized_shortcodes[] = 'columns';
-	$non_texturized_shortcodes[] = 'column';
-	$non_texturized_shortcodes[] = 'testimonial';
-	$non_texturized_shortcodes[] = 'contentbox';
-	return $non_texturized_shortcodes;
-});
-
-function simplecommerce_parse_markdown( $content ) {
-	if ( class_exists( 'Jetpack' ) && Jetpack::is_module_active( 'markdown' ) ) {
-		return WPCom_Markdown::get_instance()->transform( $content, array( 'unslash' => false ) );
-	}
-
-	return $content;
-}
-
-function simplecommerce_shortcode_columns( $attrs, $content = '' ) {
-	$column_count = 0;
-	$column_regex = '/\[column\]/s';
-	// Detect the number of columns specified in the content so that we know how wide to make each one.
-	if ( preg_match_all( $column_regex, $content, $matches ) ) {
-		$column_count = count($matches[0]);
-	}
-
-	$content = preg_replace( $column_regex, "[column total_count='$column_count']", $content );
-
-	if ( $column_count == 0 ) {
-		return $content;
-	}
-
-	return "<div class='u-cf'>" . do_shortcode( $content ) . "</div>";
-}
-
-function simplecommerce_shortcode_column( $attrs, $content = '' ) {
-	$parsed_attrs = shortcode_atts( array(
-		'total_count' => 1
-	), $attrs );
-	$css_class = '';
-	switch( $parsed_attrs['total_count'] ) {
-		case '2':
-			$css_class = 'one-half';
-			break;
-		case '3':
-			$css_class = 'one-third';
-			break;
-	}
-	return "<div class='nested column $css_class'>" . do_shortcode( simplecommerce_parse_markdown( $content ) ) . "</div>";
-}
-
-function simplecommerce_shortcode_testimonial( $attrs, $content = '' ) {
-	$parsed_attrs = shortcode_atts( array(
-		'name' => '',
-		'url' => '',
-		'image_url' => ''
-	), $attrs );
-
-	$cite = '';
-	$cite_style = '';
-
-	if ( !empty( $parsed_attrs['name'] ) ) {
-		$cite = $parsed_attrs['name'];
-
-		if ( !empty( $parsed_attrs['image_url'] ) ) {
-			$cite = $cite . "<img src='" . $parsed_attrs['image_url'] . "' alt='" . $parsed_attrs['name'] . "' class='cite_img' />";
-		} else {
-			// Ick. Have to do an inline style on the cite to get margin right in the absence of an image.
-			// Note intentional leading space.
-			$cite_style = " style='margin-right: 0;'";
-		}
-
-		if ( !empty( $parsed_attrs['url'] ) ) {
-			$cite = "<a href='" . $parsed_attrs['url'] . "'>$cite</a>";
-		}
-
-		$cite = "<cite$cite_style>$cite</cite>";
-	}
-
-	return "<blockquote class='testimonial'>" . do_shortcode( simplecommerce_parse_markdown( $content ) ) . $cite . "</blockquote>";
-}
-
-function simplecommerce_shortcode_toggle( $attrs, $content = '' ) {
-	// The ID of each toggle-able must be unique per request for the CSS styling to work.
-	// This is not threadsafe. :P
-	static $simplecommerce_toggle_id = 0;
-	$parsed_attrs = shortcode_atts( array(
-		'title' => '',
-		'initial_state' => 'closed'
-	), $attrs );
-
-	$id = ++$simplecommerce_toggle_id;
-
-	$check_state = $parsed_attrs['initial_state'] == 'open' ? 'checked="checked"' : '';
-
-	return "<style type='text/css'>#sc-toggle-chk-$id { display: none; } #sc-toggle-$id { display: none; } #sc-toggle-chk-$id:checked ~ #sc-toggle-$id { display: block; }</style>" .
-			"<div class='toggle-container'>" .
-			"<input type='checkbox' id='sc-toggle-chk-$id' $check_state />" .
-			"<label class='toggle noselect' for='sc-toggle-chk-$id'>" .
-				"<i class='fa fa-angle-double-down fa-lg collapsed'></i>" .
-				"<i class='fa fa-angle-double-up fa-lg expanded'></i>" .
-				 $parsed_attrs['title'] . "</label>" .
-			"<div id='sc-toggle-$id' class='toggle-content' markdown='1'>" . simplecommerce_parse_markdown( $content ) . "</div>" .
-			"</div>"; // .toggle-container
-
-}
-
-function simplecommerce_shortcode_contentbox( $attrs, $content = '' ) {
-	$parsed_attrs = shortcode_atts( array(
-		'align' => ''
-	), $attrs );
-	return "<aside class='content-box " . $parsed_attrs['align'] . "'>" . do_shortcode( simplecommerce_parse_markdown( $content ) ) . "</aside>";
-}
+require_once( 'functions-shortcodes.php' );
 
 // ==========================================================
 // Theme Settings (Customizer)
